+++ src/main/java/com/example/iotl/controller/StockController.java
@@ -1,9 +1,8 @@
 package com.example.iotl.controller;
 
 import com.example.iotl.dto.*;
+import com.example.iotl.dto.stocks.*;
 import com.example.iotl.entity.StockDetail;
-import com.example.iotl.entity.Stocks;
-import com.example.iotl.repository.StockInfoRepository;
 import com.example.iotl.service.StockService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
