+++ src/main/java/com/example/iotl/dto/dashboard/UserInvestmentSummaryDto.java
@@ -1,0 +1,4 @@
+package com.example.iotl.dto.dashboard;
+
+public class UserInvestmentSummaryDto {
+}
